Update README.md
There are two types of users who work on this system, the first user is the administrator and he can access the interface for entering event data or adding an event with all its details, In the login window, if the user wants to enter as admin, he has to enter in admin’s ID and admin’s password. Otherwise, he can enter as a normal user to book a ticket through enter his ID and password.
while the second user who is dedicated to the student can book a ticket by viewing a list of available events, the student can choose an activity From the list and then pressing the BOOK button. The same interface contains another tab through which the user can view all the tickets that he has booked. The system is distinguished by the presence of a specific mechanism that manages the ticket reservation process. Each user can book one ticket for the same event. The system also manages the reservation process within the scope of the number of seats available and specified in the capacity of the event.